third_party/WebKit/LayoutTests/fast/loader/javascript-url-in-embed-expected.txt

CONSOLE MESSAGE: line 9: Uncaught TypeError: Cannot set property 'src' of null
This tests that javascript: urls in an embed tag do not get executed.
